Move StorageUtils to ...beacon.chain.storage package (chain module).
